Fixed all unit tests and added (more) tests for out/ref parameters.

+++ Source/Smocks.Tests/FuncTests.cs
@@ -120,7 +120,7 @@ namespace Smocks.Tests
             Assert.AreEqual(expected, value);
         }
 
-        [TestCase((int)1)]
+        [TestCase(1)]
         public void Run_IntFunc_ReturnsCorrectValue(int expected)
         {
             int value = Smock.Run(context =>
